Fix: Error: Callback was already called
In certain odd cases, like a website returning a 302 and/or closing the
connection after a HEAD (I'm not 100% sure), linkCheck will call our
callback twice. `async` is not amused and will throw:

> Error: Callback was already called

A campaign of ad-hoc `console.log()`s suggest this is a bug in needle, a
library used by link-check.

Fixes #222 and #473